backend/requirements.txt
apimatic-core==0.2.15
apimatic-core-interfaces==0.1.5
apimatic-requests-client-adapter==0.1.6
asgiref==3.8.1
CacheControl==0.12.14
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
jsonpickle==3.0.4
jsonpointer==2.4
msgpack==1.1.0
paypal-server-sdk==0.5.2
pillow==11.1.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
